수동 로또 구입 view 구현 및 로또 번호 split util 분리
